requirements.txt
Flask==2.3.2
matplotlib==3.7.1
numpy==1.23.5
paddleocr==2.6.1.3
Pillow==10.2.0
PyMuPDF==1.20.2
pytesseract==0.3.10
